updated some controller code to allow first call from angular to pass ModelState check

+++ PRSProject/Controllers/ProductsController.cs
@@ -99,6 +99,7 @@ namespace PRSProject.Controllers
 
         public ActionResult Remove([FromBody] Product product)
         {
+            if (product.Name == null) return new EmptyResult();
             Product tempProduct = db.Products.Find(product.Id);
             db.Products.Remove(tempProduct);
             return TrySave("removed.");
